fix: restore Evaluate job build against latest Plutus master
The weekly compatibility job could no longer resolve a build plan for
run-script-evaluations: current Plutus master requires cardano-base >=0.1.5
and aeson >=2.3, which the stale index-states and flake inputs did not admit.

- Bump the CHaP and Hackage index-states to match Plutus master's cabal.project
  (CHaP 2026-06-18 exposes cardano-base 0.1.5.0).
- Update the hackage.nix and CHaP flake inputs so haskell.nix's snapshots cover
  those dates.
- Add the allow-newer/constraints Plutus master uses for the aeson >=2.3 cascade
  (deriving-aeson, microstache, setup.optparse-applicative) plus
  postgresql-simple:aeson for our own dependency.
- Adapt FromField SatInt to the new Int64-backed SatInt
  (unsafeToSatInt :: Int64 -> SatInt).

Verified locally: both executables build against Plutus master (266f6a0) under
GHC 9.6.6; fourmolu and hlint are clean.

+++ run-script-evaluations/run-evaluations/Database/PostgreSQL/Simple/Orphans.hs
@@ -3,6 +3,7 @@
 module Database.PostgreSQL.Simple.Orphans where
 
 import Control.Exception (Exception)
+import Data.Int (Int64)
 import Data.SatInt (unsafeToSatInt)
 import Database.PostgreSQL.Simple.FromField (FromField (..), conversionError)
 import PlutusLedgerApi.Common (
@@ -31,4 +32,4 @@ deriving newtype instance FromField ExMemory
 deriving newtype instance FromField ExCPU
 
 instance FromField SatInt where
-  fromField f mdata = unsafeToSatInt <$> fromField @Int f mdata
+  fromField f mdata = unsafeToSatInt <$> fromField @Int64 f mdata
